pyproject.toml
[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.uv-dynamic-versioning]
vcs = "git"
style = "pep440"
bump = true

[project]
name = "omnicoreagent"
dynamic = ["version"]
description = "OmniCoreAgent is a powerful Python AI Agent framework for building autonomous AI agents that think, reason, and execute complex tasks. Production-ready agents that use tools, manage memory, coordinate workflows, and handle real-world business logic."
readme = "README.md"
authors = [{ name = "Abiola Adeshina", email = "abiolaadedayo1993@gmail.com" }]
requires-python = ">=3.10"

keywords = ["git", "mcp", "llm", "automation", "agent", "ai", "framework"]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "colorlog>=6.9.0",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "websockets>=15.0.1",
    "httpx>=0.26.0",
    "httpx-sse>=0.4.0",
    "pydantic[email]>=2.6.0",
    "anyio>=4.2.0",
    "redis>=5.2.1",
    "python-decouple>=3.8",
    "fastapi>=0.115.12",
    "python-multipart>=0.0.20",
    "colorama>=0.4.6",
    "mcp[cli]>=1.9.1",
    "sqlalchemy>=2.0.0",
    "tzlocal>=5.2",
    "psycopg2-binary>=2.9.10",
    "apscheduler>=3.11.0",
    "psutil>=7.0.0",
    "pymupdf>=1.26.3",
    "litellm>=1.75.2",
    "fastapi-sso>=0.18.0",
    "opik>=1.8.19",
    "cryptography>=45.0.6",
    "motor>=3.7.1",
    "pymongo>=4.15.1",
    "tiktoken>=0.8.0",
]

[project.scripts]
omnicoreagent= "omnicoreagent.omni_agent.agent:OmniCoreAgent"

[project.urls]
Repository = "https://github.com/omnirexflora-labs/omnicoreagent"
Issues = "https://github.com/omnirexflora-labs/omnicoreagent/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/omnicoreagent"]

[tool.hatch.build]
packages = ["src/omnicoreagent"]
include = [

    "LICENSE",
    "README.md"
]

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "asyncio: marks tests that use asyncio",
    "OpenAIIntegration: marks tests that require an OpenAI API key",
]

[tool.uv.sources]
omnicoreagent = { workspace = true }

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "ruff>=0.11.7",
    "hatch>=1.14.1",
    "twine>=6.1.0",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-git-revision-date-localized-plugin>=1.4.7",
    "mkdocs-material>=9.6.14",
    "mkdocs-minify-plugin>=0.8.0",
    "pymdown-extensions>=10.16",
]
